Implement tagged signing traits for Schnorr and BLS keys
Implement the tagged signing traits `TaggedPublicKey` and
`TaggedKeyPair` for `BLS` and `Ed25519`  keypairs/keys.
