+++ bls/Cargo.toml
@@ -34,7 +34,7 @@ ark-serialize = "0.4"
 nimiq-hash = { workspace = true }
 nimiq-hash_derive = { workspace = true }
 nimiq-serde = { workspace = true, optional = true }
-nimiq-utils = { workspace = true, features = ["key-rng"] }
+nimiq-utils = { workspace = true, features = ["key-rng", "tagged-signing"] }
 
 [dev-dependencies]
 nimiq-test-log = { workspace = true }
